feat: add EmailOTPDialog component for email verification
- Introduced `EmailOTPDialog` to handle email OTP verification, including countdown timer and resend functionality.
- Integrated the dialog into the `useOneKeyAuth` hook for streamlined email OTP sending and confirmation.
- Updated various components to utilize the new authentication structure, replacing references to the deprecated `useLoginOneKeyId` hook.
- Enhanced user experience with improved error handling and toast notifications for verification feedback.
