+++ packages/kit/src/views/Prime/components/PrimeDeviceLogoutAlertDialog.tsx
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import { useIntl } from 'react-intl';
 
 import { Dialog, SizableText, Stack } from '@onekeyhq/components';
-import { usePrimeAuthV2 } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
+import { useOneKeyAuth } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import { defaultLogger } from '@onekeyhq/shared/src/logger/logger';
 
@@ -12,7 +12,7 @@ import { useDebounce } from '../../../hooks/useDebounce';
 export function PrimeDeviceLogoutAlertDialog() {
   const intl = useIntl();
 
-  const { logout } = usePrimeAuthV2();
+  const { logout } = useOneKeyAuth();
 
   const logoutDebounced = useDebounce(logout, 600, {
     leading: false,
