+++ packages/kit/src/views/Prime/components/PrimePurchaseDialog/PrimePurchaseDialog.tsx
@@ -10,7 +10,7 @@ import {
   YStack,
 } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { usePrimeAuthV2 } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
+import { useOneKeyAuth } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
 import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
 import googlePlayService from '@onekeyhq/shared/src/googlePlayService/googlePlayService';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
@@ -35,7 +35,7 @@ export function usePrimePurchaseCallback({
     purchasePackageWeb,
     getPackagesWeb,
   } = usePrimePayment();
-  const { user } = usePrimeAuthV2();
+  const { user } = useOneKeyAuth();
   const intl = useIntl();
 
   const purchaseByWebview = usePurchasePackageWebview();
@@ -166,7 +166,7 @@ export const PrimePurchaseDialog = (props: {
 }) => {
   const { onPurchase, featureName } = props;
   const intl = useIntl();
-  const { user } = usePrimeAuthV2();
+  const { user } = useOneKeyAuth();
   const [selectedSubscriptionPeriod, setSelectedSubscriptionPeriod] =
     useState<ISubscriptionPeriod>('P1Y');
