+++ packages/kit/src/views/Prime/hooks/usePrimeAvailable.tsx
@@ -1,13 +1,13 @@
 import { useMemo } from 'react';
 
-import { usePrimeAuthV2 } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
+import { useOneKeyAuth } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
 import { useDevSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms/devSettings';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
 export function usePrimeAvailable() {
   const [devSettings] = useDevSettingsPersistAtom();
 
-  const { user } = usePrimeAuthV2();
+  const { user } = useOneKeyAuth();
 
   const isPrimeAvailable = useMemo(() => {
     if (devSettings.enabled && devSettings.settings?.showPrimeTest) {
