+++ packages/kit/src/views/Prime/pages/PrimeDashboard/PrimeDashboard.tsx
@@ -18,7 +18,7 @@ import {
   useSafeAreaInsets,
 } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { usePrimeAuthV2 } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
+import { useOneKeyAuth } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
 import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
 import type { IOneKeyError } from '@onekeyhq/shared/src/errors/types/errorTypes';
@@ -81,16 +81,21 @@ export default function PrimeDashboard({
   const { fromFeature } = route.params || {};
   // const isReady = false;
   const {
+    isReady: isAuthReady,
     user,
     isLoggedIn,
     isPrimeSubscriptionActive,
     supabaseUser,
     isSupabaseLoggedIn,
     // logout,
-  } = usePrimeAuthV2();
+  } = useOneKeyAuth();
 
-  const { isReady, getPackagesNative, restorePurchases, getPackagesWeb } =
-    usePrimePayment();
+  const {
+    isReady: isPurchaseReady,
+    getPackagesNative,
+    restorePurchases,
+    getPackagesWeb,
+  } = usePrimePayment();
 
   const [selectedSubscriptionPeriod, setSelectedSubscriptionPeriod] =
     useState<ISubscriptionPeriod>('P1Y');
@@ -115,7 +120,7 @@ export default function PrimeDashboard({
   useEffect(() => {
     const fn = async () => {
       // isFocused won't be triggered when Login Dialog is open or closed
-      if (isFocused) {
+      if (isFocused && isAuthReady) {
         await timerUtils.wait(600);
         if (!isFocusedRef.current) {
           // may be blurred when auto navigate to Device Limit Page
@@ -127,7 +132,7 @@ export default function PrimeDashboard({
       }
     };
     void fn();
-  }, [isFocused]);
+  }, [isFocused, isAuthReady]);
 
   const shouldShowConfirmButton = useMemo(() => {
     if (!isLoggedIn || !isPrimeSubscriptionActive) {
@@ -156,8 +161,8 @@ export default function PrimeDashboard({
   // }, []);
 
   const { result: webPackages } = usePromiseResult(async () => {
-    if (isReady) {
-      console.log('getPackagesWeb2__isReady', isReady);
+    if (isPurchaseReady) {
+      console.log('getPackagesWeb2__isReady', isPurchaseReady);
       const shouldPolyfillRandomUUIDTemporarily =
         !globalThis?.crypto?.randomUUID && platformEnv.isNativeAndroid;
       if (shouldPolyfillRandomUUIDTemporarily) {
@@ -179,12 +184,13 @@ export default function PrimeDashboard({
         }
       }
     }
-  }, [getPackagesWeb2, isReady]);
+  }, [getPackagesWeb2, isPurchaseReady]);
 
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { result: sdkPackages, isLoading: isPackagesLoading } =
     usePromiseResult(
       async () => {
-        if (!shouldShowSubscriptionPlans || !isReady) {
+        if (!shouldShowSubscriptionPlans || !isPurchaseReady) {
           return [];
         }
 
@@ -260,7 +266,7 @@ export default function PrimeDashboard({
         intl,
         getPackagesNative,
         getPackagesWeb,
-        isReady,
+        isPurchaseReady,
         shouldShowSubscriptionPlans,
         user?.onekeyUserId,
       ],
@@ -326,6 +332,7 @@ export default function PrimeDashboard({
   const isLoggedInMaybe =
     isSupabaseLoggedIn ||
     supabaseUser?.id ||
+    user?.onekeyUserId ||
     user?.isLoggedIn ||
     user?.isLoggedInOnServer ||
     isLoggedIn;
@@ -386,7 +393,7 @@ export default function PrimeDashboard({
               </Stack>
             ) : null}
 
-            {isReady ? (
+            {isPurchaseReady ? (
               <PrimeBenefitsList
                 selectedSubscriptionPeriod={selectedSubscriptionPeriod}
                 networkId={route.params?.networkId}
