+++ packages/kit/src/views/Prime/pages/PrimeDashboard/PrimeDebugPanel.tsx
@@ -9,9 +9,8 @@ import {
   XStack,
 } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { usePrimeAuthV2 } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
+import { useOneKeyAuth } from '@onekeyhq/kit/src/components/OneKeyAuth/useOneKeyAuth';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
-import { useLoginOneKeyId } from '@onekeyhq/kit/src/hooks/useLoginOneKeyId';
 import { usePrimePayment } from '@onekeyhq/kit/src/views/Prime/hooks/usePrimePayment';
 import { usePrimePersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { OneKeyLocalError } from '@onekeyhq/shared/src/errors';
@@ -188,13 +187,12 @@ export function PrimeDebugPanel({
 }: {
   shouldShowConfirmButton: boolean;
 }) {
-  const { getAccessToken, logout, isReady, isSupabaseLoggedIn } =
-    usePrimeAuthV2();
+  const { getAccessToken, logout, isReady, isSupabaseLoggedIn, loginOneKeyId } =
+    useOneKeyAuth();
   const { getCustomerInfo, getPackagesNative, getPackagesWeb } =
     usePrimePayment();
   const navigation = useAppNavigation();
   const [isHidden, setIsHidden] = useState(false);
-  const { loginOneKeyId } = useLoginOneKeyId();
   const [primePersistAtomData] = usePrimePersistAtom();
   const { purchaseByWebview } = usePrimePurchaseCallback();
